+++ druid/src/core.rs
@@ -714,6 +714,10 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 }
                 true
             }
+            Event::WindowScale(_) => {
+                self.state.needs_layout = true;
+                true
+            }
             Event::WindowSize(_) => {
                 self.state.needs_layout = true;
                 ctx.is_root